[fleet] Add `on_exit` handler to the Pty
We want to track the child process' termination and retrieve its exit
code.

To achieve that, a new parameter in the Pty creation API was introduced,
which allows to provide a callback for the event of the child process
termination.

The callback receives a single parameter of the `ExitStatus` type, which
is similar to the type of the same name in Fleet.